Use rgba colors to avoid Safari rendering issue
Our error pages include a repeated "watermark" background image showing
the HTTP status code, so that even if the error message is omitted or
obscured from a screenshot, we can still discern the status code.

Unfortunately Safari's SVG rendering highlights the edges of the
watermark, making it very visible and very ugly and leaving the
underlying page contents unreadable. (ht to @mmcwatters for flagging.)

The problem is rectified by instead using rgba colors: the alpha
is then rendered to blend with the white background.

I've also made the watermarks much subtler in this version. They're
only intended as a fallback to make sure our friendly error messages
don't get in the way of debugging; there's no need for them to be
discernible with the naked eye.